fix: rename entities utilities file
