+++ src/entitiesUtilities.ts
